add check status in posts and friends

+++ src/main/java/com/npxception/demo/dao/PostgreSQLPostDaoImpl.java
@@ -62,6 +62,7 @@ public class PostgreSQLPostDaoImpl implements PostDao {
 
   @Override
   public Collection<Post> getPostsByUserFromGroup(int userId, int groupId) {
+    checkMember(userId, groupId);
     final String sql1 = "SELECT * FROM users WHERE userid = ? ";
     User user = jdbcTemplate.queryForObject(sql1, new UserRowMapper(), userId);
     String author = user.getFirstName() + " " + user.getLastName();
@@ -69,6 +70,21 @@ public class PostgreSQLPostDaoImpl implements PostDao {
         "AND visibility = ? ORDER BY time";
     List<Post> posts = jdbcTemplate.query(sql2, new PostRowMapper(), new Object[]{author, groupId});
     return posts;
+  }
+
+  @Override
+  public Collection<Post> getPostsByUserFromGroupName(int id, String name) {
+    int groupid = jdbcTemplate.queryForObject("SELECT groupid FROM groups WHERE groupname = ?"
+        , new Object[]{name}, Integer.class);
+    return getPostsByUserFromGroup(id, groupid);
+  }
+
+  public void checkMember(int userId, int groupId) {
+    final String sql = "SELECT * FROM membership WHERE memberid = ? AND groupid = ?";
+    Object object = jdbcTemplate.queryForObject(sql, new Object[]{userId, groupId}, Object.class);
+    if (object == null) {
+      throw new AuthenticationException(userId);
+    }
   }
 
 //  @Override
@@ -100,9 +116,28 @@ public class PostgreSQLPostDaoImpl implements PostDao {
   }
 
   @Override
-  public void removePostsById(int id) {
+  public void removePostsById(int id1, int id) {
+    checkAuthor(id1, id);
     final String sql = "DELETE FROM posts WHERE id = ?";
     jdbcTemplate.update(sql, id);
+  }
+
+  //to see if the post(id) is written by the user(id1)
+  public void checkAuthor(int id1, int id) {
+    String authorfirst = jdbcTemplate.queryForObject
+        ("SELECT authorfirst FROM posts WHERE id = ?",
+        new Object[]{id},
+        String.class);
+    String authorlast = jdbcTemplate.queryForObject
+        ("SELECT authorlast FROM posts WHERE id = ?",
+            new Object[]{id},
+            String.class);
+    String sql = "SELECT * FROM users WHERE firstname = ? AND lastname = ?";
+    User user = jdbcTemplate.queryForObject(sql, new UserRowMapper(),
+        new Object[]{authorfirst, authorlast});
+    if (user == null){
+      throw new AuthenticationException(id1);
+    }
   }
 
   @Override
@@ -197,20 +232,6 @@ public class PostgreSQLPostDaoImpl implements PostDao {
       post.setTime(resultSet.getInt("time"));
       post.setVisibility(resultSet.getInt("visibility"));
       return post;
-    }
-  }
-
-  public void checkUser(int id){
-    String sql0 = "SELECT * FROM users WHERE userid = ?";
-    User user = jdbcTemplate.queryForObject(sql0, new Object[]{id}, new UserRowMapper());
-    String email = user.getEmail();
-    String password = user.getPassword();
-    String sql1 = "SELECT email FROM loginfo";
-    String sql2 = "SELECT password FROM loginfo";
-    String loginEmail = jdbcTemplate.queryForObject(sql1, new Object[]{}, String.class);
-    String loginPass = jdbcTemplate.queryForObject(sql2, new Object[]{}, String.class);
-    if ((!email.equals(loginEmail)) || (!password.equals(loginPass))){
-      throw new AuthenticationException(id);
     }
   }
 }
